Fix mutability issues
The actual input variable had a mutable value. Puzzle 05 was the
first one to nastily mutate the input object it got, causing un-
wanted interaction between the two puzzles. We "solve" this with
simply cloning (in an inefficient but straightforward manner) the
input object before throwing it at a test.

Should've used a language that warns/prevent of this kind of thing,
perhaps :D