Refactor day 23
This is the optimized solution, where we realize that the inner
loops are looking for non-primes to increase the H register.

In fact, during the past hours, I did try this solution, but had
the logic for prime-checking inverted.

See the diff for this commit to check how this high level version
came from the low level one.